q3map2/picomodel: remove some warnings
imported from netradiant, initial commit message:

> q3map2 is now waring free

+++ tools/quake3/common/cmdlib.c
@@ -188,7 +188,7 @@ void SetQdirFromPath( const char *path ){
 			}
 			strncpy( qdir, path, c + len + count - path );
 			Sys_Printf( "qdir: %s\n", qdir );
-			for ( i = 0; i < strlen( qdir ); i++ )
+			for ( i = 0; i < (int) strlen( qdir ); i++ )
 			{
 				if ( qdir[i] == '\\' ) {
 					qdir[i] = '/';
@@ -201,7 +201,7 @@ void SetQdirFromPath( const char *path ){
 				if ( *c == '/' || *c == '\\' ) {
 					strncpy( gamedir, path, c + 1 - path );
 
-					for ( i = 0; i < strlen( gamedir ); i++ )
+					for ( i = 0; i < (int) strlen( gamedir ); i++ )
 					{
 						if ( gamedir[i] == '\\' ) {
 							gamedir[i] = '/';
